:sparkles: feat: add active class to menu item

+++ app/Livewire/Admin/Categories.php
@@ -5,15 +5,19 @@ namespace App\Livewire\Admin;
 use Livewire\Component;
 use App\Models\ParentCategory;
 use App\Models\Category;
-
+use Livewire\WithPagination;
 
 class Categories extends Component
 {
+    use WithPagination;
     public $isUpdateParentCategoryMode = false;
     public $pcategory_id, $pcategory_name;
 
     public $isUpdateCategoryMode = false;
     public $category_id, $parent = 0, $category_name;
+
+    public $pcategoriesPerPage = 5;
+    public $categoriesPerPage = 10;
 
     protected $listeners = [
         'updateParentCategoryOrdering',
@@ -117,6 +121,14 @@ class Categories extends Component
     public function deleteParentCategoryAction($id)
     {
         $pcategory = ParentCategory::findOrFail($id);
+
+        if($pcategory->children->count() > 0){
+            foreach($pcategory->children as $category){
+                Category::where('id', $category->id)->update(['parent' => 0]);
+            }
+            return;
+        }
+
         $delete = $pcategory->delete();
 
         if($delete){
@@ -224,8 +236,8 @@ class Categories extends Component
     public function render()
     {
         return view('livewire.admin.categories', [
-            'pcategories' => ParentCategory::orderBy('ordering','asc')->get(),
-            'categories' => Category::orderBy('name','asc')->get()
+            'pcategories' => ParentCategory::orderBy('ordering','asc')->paginate($this->pcategoriesPerPage,['*'],'pcat_page'),
+            'categories' => Category::orderBy('ordering','asc')->paginate($this->categoriesPerPage, ['*'], 'cat_page')
         ]);
     }
 }
